Make dora-rerun pip installable and python-abi3
