+++ node-hub/dora-rerun/Cargo.toml
@@ -9,6 +9,10 @@ repository.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+default = []
+python = ["pyo3"]
+
 [dependencies]
 dora-node-api = { workspace = true, features = ["tracing"] }
 eyre = "0.6.8"
@@ -16,3 +20,13 @@ tokio = { version = "1.24.2", features = ["rt"] }
 rerun = { version = "0.19.0", features = ["web_viewer", "image"] }
 ndarray = "0.15.6"
 k = "0.32"
+pyo3 = { workspace = true, features = [
+    "extension-module",
+    "abi3",
+], optional = true }
+
+
+[lib]
+name = "dora_rerun"
+path = "src/lib.rs"
+crate-type = ["rlib", "cdylib"]
